+++ ThunderStore/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+- Added FunoPlush Valuable
+- Added SCP-999 Plush Valuable
+- Added Candy Valuables
+- Lowered the durability of the nuke
+
 ## 1.0.1
 - Changed model for nuke
 - Changed how nuke works so it doesnt lag the game